third_party/WebKit/LayoutTests/http/tests/security/suborigins/crossorigin/suborigin-cors-preflight-xhr-failure-output-expected.txt

CONSOLE ERROR: Failed to load http://127.0.0.1:8000/security/resources/cors-script.php?cors=false&with-preflight: Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS Custom headers causes preflight failure
Harness: the test ran to completion.
